Fixing errors

+++ app/Interfaces/PaymentPickupable.php
@@ -23,7 +23,8 @@ interface PaymentPickupable
 	/**
 	 * Set added paid amount after the payment
 	 * 
+	 * @param float  $amount
 	 * @return void
 	 */
-	public function setAddedPaidAmountAttribute(double $amount);
+	public function setAddedPaidAmountAttribute(float $amount);
 }
